+++ changelog.md
@@ -1,5 +1,6 @@
 ## 2.4.0
 - [FEA][COR] carousel view model
+- [FEA][AND] recipe carousel component
 - [FIX][IOS] Infinite loading on recipe card when using a template
 - [FIX][IOS] Add missing templates